Update nonce constants to be uint64

+++ packages/contracts-communication/test/InterchainClientV1.Dst.t.sol
@@ -28,10 +28,10 @@ import {InterchainDBMock} from "./mocks/InterchainDBMock.sol";
 /// 4. Execute the transaction by passing the message to the recipient.
 /// 5. Mark transaction as executed and emit an event.
 contract InterchainClientV1DestinationTest is InterchainClientV1BaseTest {
-    uint256 public constant MOCK_DB_NONCE = 444;
+    uint64 public constant MOCK_DB_NONCE = 444;
     uint64 public constant MOCK_ENTRY_INDEX = 4;
 
-    uint256 public constant MOCK_LOCAL_DB_NONCE = 123;
+    uint64 public constant MOCK_LOCAL_DB_NONCE = 123;
     uint64 public constant MOCK_LOCAL_ENTRY_INDEX = 5;
 
     uint256 public constant MOCK_GAS_LIMIT = 100_000;
